# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TriageDesk API
  description: "Help desk ticketing for a small IT team: tickets, queues, SLA timers\
    \ that count business hours, an audit trail and reporting. Errors are RFC 9457\
    \ Problem Details with a stable `type` per error and a `correlation_id` on every\
    \ response."
  license:
    name: MIT
    url: https://github.com/gavinfecko/triagedesk-api/blob/main/LICENSE
  version: 0.1.0-SNAPSHOT
...
servers:
- url: /
  description: This deployment
security:
- bearerAuth: []
...
paths:
...
  /api/v1/tickets:
    get:
      tags:
      - Tickets
      summary: "List tickets, newest first: your own as a requester, all of them as\
        \ an agent or admin"
      operationId: list_1
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 25
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseTicketSummary"
        "400":
          description: Validation failed or the request is malformed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: The caller's role does not allow this
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: The resource does not exist or is not visible to the caller
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: A business rule refused the request (see the type slug)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Unexpected error; quote the correlation id
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    post:
      tags:
      - Tickets
      summary: Open a ticket
      description: Agents and admins may set requester_id (a ticket phoned in) and
        queue_id. Requesters always open their own ticket in the category's queue;
        staff-only fields they send are ignored and listed in warnings.
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketView"
        "400":
          description: Validation failed or the request is malformed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: The caller's role does not allow this
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: The resource does not exist or is not visible to the caller
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: A business rule refused the request (see the type slug)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Unexpected error; quote the correlation id
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/v1/tickets/{key}/transitions:
    post:
      tags:
      - Tickets
      summary: Move a ticket to another status
      description: "NEW→OPEN (staff), OPEN→PENDING (staff, comment required), PENDING→\
        OPEN (anyone), OPEN|PENDING→RESOLVED (staff, comment required), RESOLVED→\
        CLOSED (requester or admin), RESOLVED→OPEN (anyone), NEW→CANCELLED (requester\
        \ or admin), OPEN|PENDING→CANCELLED (admin). Anything else is 409 ticket-state-conflict.\
        \ A comment is stored as a public reply."
      operationId: transition
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketView"
        "400":
          description: Validation failed or the request is malformed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: The caller's role does not allow this
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: The resource does not exist or is not visible to the caller
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: A business rule refused the request (see the type slug)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Unexpected error; quote the correlation id
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/v1/tickets/{key}/queue:
    post:
      tags:
      - Tickets
      summary: Move a ticket to another queue (staff)
      operationId: moveToQueue
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueueRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketView"
        "400":
          description: Validation failed or the request is malformed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: The caller's role does not allow this
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: The resource does not exist or is not visible to the caller
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: A business rule refused the request (see the type slug)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Unexpected error; quote the correlation id
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
  /api/v1/tickets/{key}/assign:
    post:
      tags:
      - Tickets
      summary: "Assign a ticket (staff): \"me\", another agent's or admin's id, or\
        \ null to unassign"
      description: Taking a NEW ticket opens it.
      operationId: assign
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketView"
        "400":
          description: Validation failed or the request is malformed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: The caller's role does not allow this
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: The resource does not exist or is not visible to the caller
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: A business rule refused the request (see the type slug)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Unexpected error; quote the correlation id
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
  /api/v1/tickets/{key}:
    get:
      tags:
      - Tickets
      summary: One ticket by its key (HD-001234); someone else's ticket is a 404 for
        requesters
      description: The response carries an ETag; send it back as If-Match when you
        PATCH.
      operationId: get
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketView"
        "400":
          description: Validation failed or the request is malformed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: The caller's role does not allow this
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: The resource does not exist or is not visible to the caller
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: A business rule refused the request (see the type slug)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Unexpected error; quote the correlation id
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    patch:
      tags:
      - Tickets
      summary: "Edit a ticket: title, description, priority, category"
      description: "Requires If-Match with the ETag you last read: missing is 428,\
        \ stale is 412. Staff may edit any open ticket; a requester may change the\
        \ title and description of their own ticket only while it is NEW, and never\
        \ the priority or category. Changing the category does not move the queue\
        \ (see /queue)."
      operationId: update_1
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      - name: If-Match
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditTicketRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketView"
        "400":
          description: Validation failed or the request is malformed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: The caller's role does not allow this
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: The resource does not exist or is not visible to the caller
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: A business rule refused the request (see the type slug)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Unexpected error; quote the correlation id
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
components:
  schemas:
...
    CreateTicketRequest:
      properties:
        title:
          type: string
          maxLength: 120
          minLength: 5
        description:
          type: string
          maxLength: 5000
          minLength: 0
        category_id:
          type: string
          format: uuid
        priority:
          type: string
          enum:
          - P1_CRITICAL
          - P2_HIGH
          - P3_MEDIUM
          - P4_LOW
        requester_id:
          type: string
          format: uuid
        queue_id:
          type: string
          format: uuid
      required:
      - category_id
      - description
      - priority
      - title
    Person:
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
    Ref:
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    TicketView:
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
          - NEW
          - OPEN
          - PENDING
          - RESOLVED
          - CLOSED
          - CANCELLED
        priority:
          type: string
          enum:
          - P1_CRITICAL
          - P2_HIGH
          - P3_MEDIUM
          - P4_LOW
        category:
          $ref: "#/components/schemas/Ref"
        queue:
          $ref: "#/components/schemas/Ref"
        requester:
          $ref: "#/components/schemas/Person"
        assignee:
          $ref: "#/components/schemas/Person"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        first_responded_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
        warnings:
          type: array
          items:
            type: string
    TransitionRequest:
      properties:
        to:
          type: string
          enum:
          - NEW
          - OPEN
          - PENDING
          - RESOLVED
          - CLOSED
          - CANCELLED
        comment:
          type: string
          maxLength: 10000
          minLength: 0
      required:
      - to
    QueueRequest:
      properties:
        queue_id:
          type: string
          format: uuid
      required:
      - queue_id
...
    AssignRequest:
      properties:
        assignee_id:
          type: string
...
    EditTicketRequest:
      properties:
        title:
          type: string
          maxLength: 120
          minLength: 5
        description:
          type: string
          maxLength: 5000
          minLength: 0
        priority:
          type: string
          enum:
          - P1_CRITICAL
          - P2_HIGH
          - P3_MEDIUM
          - P4_LOW
        category_id:
          type: string
          format: uuid
    PageMeta:
      properties:
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total_elements:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int32
...
    PageResponseTicketSummary:
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TicketSummary"
        page:
          $ref: "#/components/schemas/PageMeta"
    TicketSummary:
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
        title:
          type: string
        status:
          type: string
          enum:
          - NEW
          - OPEN
          - PENDING
          - RESOLVED
          - CLOSED
          - CANCELLED
        priority:
          type: string
          enum:
          - P1_CRITICAL
          - P2_HIGH
          - P3_MEDIUM
          - P4_LOW
        category:
          type: string
        queue:
          type: string
        requester:
          $ref: "#/components/schemas/Person"
        assignee:
          $ref: "#/components/schemas/Person"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ProblemDetail:
      type: object
      description: RFC 9457 Problem Details. `type` is a stable slug under /problems/
        that clients can switch on.
      properties:
        type:
          type: string
          example: /problems/validation
        title:
          type: string
          example: Validation failed
        status:
          type: integer
          format: int32
          example: 400
        detail:
          type: string
          example: 2 fields failed validation
        instance:
          type: string
          example: /api/v1/tickets
        correlation_id:
          type: string
          example: 8b1f2c0e-4d3a-4f0e-9a6b-2c7d1e5f9a10
        errors:
          type: array
          description: Present on validation problems only
          items:
            type: object
            properties:
              field:
                type: string
                example: title
              message:
                type: string
                example: size must be between 5 and 120
      required:
      - correlation_id
      - status
      - title
      - type
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT